refactor(auth): split auth.ts into jwt/token-storage/oauth modules (#5)
